makersboard/requirements.txt
Babel==2.3.3
Django==1.9.5
Pillow==3.2.0
django-appconf==1.0.1
django-phonenumber-field==1.1.0
mock==2.0.0
pbr==1.9.1
phonenumberslite==7.3.0
pytz==2016.3
six==1.10.0
